core: Add ipsec offload support
This is the initial implementation of the ipsec offload handling.
The change adds the xfrmdev ops which can used by the physical
driver to set the relevant handlers.

Statistics are exposed through ethtool interface for debugging
purposes.

Note that ethernet and ipsec cannot be supported concurrently
as of now.

CRs-Fixed: 3607833
Change-Id: Ic624a31fe0e974a5ce40df81ebcf49677f030d95
Signed-off-by: Subash Abhinov Kasiviswanathan <quic_subashab@quicinc.com>