core: Change ipsec offload packet detection scheme
Apparently messing around with the skb control buffer for packet offload
has some unexpected consequences especially for packets with multiple
segments. This caused the single segment packets to go through the
tunnel while the other multiple segment packets were skipping the
tunnel leading to TCP transmissions. The receiver was smart enough to
figure out the difference in the path of these packets and was dropping
the multiple segment packets leading to retransmissions and poor TCP
throughput.

A workaround is added now by adding a field in the main sk_buff struct
itself to identify the skbs meant for packet offload.

CRs-Fixed: 3874201
Change-Id: I92816e3a5c4ebc04920c9d0bf39b02b65648f590
Signed-off-by: Subash Abhinov Kasiviswanathan <quic_subashab@quicinc.com>