core: Update queueing logic for packets without full sockets
The CGROUP_INET_EGRESS hook has checks to bypass the filter in case
the packet doesn't have a socket or a full socket. This results in
these packets ending up in the default queue causing it to be stuck
for a while if flow is disabled.

As there is no alternate filter installed at this point, allow
the packets with no socket or mini socket through.

CRs-Fixed: 3892186
Change-Id: I165cd3899ba6fb09b08e2c462a7b19b945f4b6f3
Signed-off-by: Subash Abhinov Kasiviswanathan <quic_subashab@quicinc.com>