rmnet_core: add new hook for rmnet_perf
rmnet_perf can now drop packets, and needs a new hook to do so safely.
The existing RX handler hook is too late in the callflow to be used, and
the exisitng ingress hook is dangerous to drop packets in. This new hook
is called just before any SHS processing on the packet.

CRs-Fixed: 4039328
Change-Id: Icd7b83b793d3256540a7907d39dadf2e47f1a00a
Signed-off-by: Sean Tranchetti <quic_stranche@quicinc.com>