core: Setting low latency bit when LLC is not present
UL prioritization was not set for the packets if the low latency channels are not present at IPA.
Adding support in vienna for setting Ul prio even if the low latency channel is not present at the ipa. We set the bit for the packets whose LLB is set, for IPA q6 to prioritize over regular channels.

Change-Id: I65b12caec951c8f16d5bd9e0c1de27fc61422f8f
Signed-off-by: Krashnika Gulati <quic_krashnik@quicinc.com>