datarmnet: Agnostically fix the trace include paths
With the advent of shared branches, it is high time to purge the awful
chain of #ifdefs and dot-dots pointing to where trace files are; what is
correct for one platform on the branch is not for another.

Instead, we do this the way kernel developers always meant for it to be
done:
  TRACE_INCLUDE_PATH is set to .
  Your include flags point to the source directory so the file can be
  found.

This is achievable with adding
CFLAGS_file_with_CREATE_TRACEPOINTS := -I$(src)
to the Makefiles and Kbuild. Now the tracefiles should be picked up
wheverever they may lie in relation to the kernel.

Change-Id: Ia0f7a7f7841b5789b202e0a5498ecbc1b5a1872d
Signed-off-by: Sean Tranchetti <sean.tranchetti@oss.qualcomm.com>